Create README.md
Greetings!
As FPX team, we are designing a crypto architecture that will be more easily used in the field of your daily life and will be valid everywhere with the great innovations we will bring to the crypto money world.

With the FPX Mobile Application, you will be able to collect your digital assets in a single wallet equipped with a high-level security circle, convert them into cash and spend them in seconds. Moreover, by earning bonus points.